pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brkraw-viewer"
version = "0.2.7"
description = "BrkRaw scan viewer plugin for brkraw CLI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "brkraw>=0.5.0rc1",
  "nibabel>=5.0",
  "pillow>=10.0",
  "ruamel.yaml>=0.17.0",
]
authors = [{name = "BrkRaw"}]
license = {text = "MIT"}
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
docs = [
  "mkdocs-material>=9.5.0",
]
test = [
  "pytest>=7.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
brkraw_viewer = ["assets/*.png", "assets/*.ico", "snippets/**/*.yaml"]

[project.entry-points."brkraw.cli"]
viewer = "brkraw_viewer.cli_viewer:register"
viewer-config = "brkraw_viewer.cli_viewer_config:register"
